Final cosmetic changes
